+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciEncryptionTest.kt
@@ -6,17 +6,14 @@ import at.asitplus.openid.TokenResponseParameters
 import at.asitplus.signum.indispensable.josef.JsonWebKey
 import at.asitplus.signum.indispensable.josef.JweEncrypted
 import at.asitplus.signum.indispensable.josef.JweHeader
-import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
+import at.asitplus.wallet.lib.agent.Holder
 import at.asitplus.wallet.lib.agent.IssuerAgent
 import at.asitplus.wallet.lib.agent.RandomSource
 import at.asitplus.wallet.lib.data.AtomicAttribute2023
 import at.asitplus.wallet.lib.data.ConstantIndex
-import at.asitplus.wallet.lib.data.VerifiableCredentialJws
+import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.PLAIN_JWT
 import at.asitplus.wallet.lib.data.rfc3986.toUri
-import at.asitplus.wallet.lib.data.vckJsonSerializer
-import at.asitplus.wallet.lib.jws.DecryptJwe
-import at.asitplus.wallet.lib.jws.DecryptJweFun
 import at.asitplus.wallet.lib.jws.EncryptJweFun
 import at.asitplus.wallet.lib.oauth2.OAuth2Client
 import at.asitplus.wallet.lib.oauth2.SimpleAuthorizationService
@@ -24,9 +21,9 @@ import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
 import at.asitplus.wallet.lib.openid.DummyUserProvider
 import at.asitplus.wallet.lib.openid.DummyOAuth2IssuerCredentialDataProvider
 import com.benasher44.uuid.uuid4
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.assertions.throwables.shouldThrowAny
 import io.kotest.core.spec.style.FunSpec
-import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.types.shouldBeInstanceOf
 
@@ -36,7 +33,6 @@ class OidvciEncryptionTest : FunSpec({
     lateinit var issuer: CredentialIssuer
     lateinit var client: WalletService
     lateinit var state: String
-    lateinit var decryptJwe: DecryptJweFun
 
     suspend fun getToken(scope: String): TokenResponseParameters {
         val authnRequest = client.oauth2Client.createAuthRequest(
@@ -69,18 +65,71 @@ class OidvciEncryptionTest : FunSpec({
                 randomSource = RandomSource.Default
             ),
             credentialSchemes = setOf(ConstantIndex.AtomicAttribute2023),
-            requireEncryption = true, // this is important, to require encryption
+            encryptionService = IssuerEncryptionService(
+                requireResponseEncryption = true, // this is important
+                decryptionKeyMaterial = EphemeralKeyWithoutCert()
+            ),
         )
         state = uuid4().toString()
-        val decryptionKeyMaterial = EphemeralKeyWithoutCert()
         client = WalletService(
-            requestEncryption = true, // this is important
-            decryptionKeyMaterial = decryptionKeyMaterial // this is important
+            encryptionService = WalletEncryptionService(
+                requestEncryption = true, // this is important
+                decryptionKeyMaterial = EphemeralKeyWithoutCert() // this is important
+            )
         )
-        decryptJwe = DecryptJwe(decryptionKeyMaterial)
     }
 
-    test("issuer fails to encrypt") {
+    test("wallet encrypts credential request and decrypts credential response") {
+        val requestOptions = WalletService.RequestOptions(ConstantIndex.AtomicAttribute2023, PLAIN_JWT)
+        val credentialFormat = client.selectSupportedCredentialFormat(requestOptions, issuer.metadata).shouldNotBeNull()
+        val scope = credentialFormat.scope.shouldNotBeNull()
+        val token = getToken(scope)
+
+        client.createEncryptedCredentialRequest(
+            tokenResponse = token,
+            metadata = issuer.metadata,
+            credentialFormat = credentialFormat,
+            clientNonce = issuer.nonce().getOrThrow().clientNonce
+        ).getOrThrow().forEach {
+            issuer.credentialEncryptedRequest(
+                authorizationHeader = token.toHttpHeaderValue(),
+                input = it,
+                credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
+            ).getOrThrow().let { credential ->
+                client.parseCredentialResponse(credential, PLAIN_JWT, ConstantIndex.AtomicAttribute2023)
+                    .getOrThrow().first().shouldBeInstanceOf<Holder.StoreCredentialInput.Vc>().apply {
+                        signedVcJws.payload.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
+                    }
+            }
+        }
+    }
+
+    test("wallet does not encrypt credential request and decrypts credential response") {
+        val requestOptions = WalletService.RequestOptions(ConstantIndex.AtomicAttribute2023, PLAIN_JWT)
+        val credentialFormat = client.selectSupportedCredentialFormat(requestOptions, issuer.metadata).shouldNotBeNull()
+        val scope = credentialFormat.scope.shouldNotBeNull()
+        val token = getToken(scope)
+
+        client.createCredentialRequest(
+            tokenResponse = token,
+            metadata = issuer.metadata,
+            credentialFormat = credentialFormat,
+            clientNonce = issuer.nonce().getOrThrow().clientNonce
+        ).getOrThrow().forEach {
+            issuer.credential(
+                authorizationHeader = token.toHttpHeaderValue(),
+                params = it,
+                credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
+            ).getOrThrow().let { credential ->
+                client.parseCredentialResponse(credential, PLAIN_JWT, ConstantIndex.AtomicAttribute2023)
+                    .getOrThrow().first().shouldBeInstanceOf<Holder.StoreCredentialInput.Vc>().apply {
+                        signedVcJws.payload.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
+                    }
+            }
+        }
+    }
+
+    test("wallet does not encrypt credential request but issuer requires this") {
         issuer = CredentialIssuer(
             authorizationService = authorizationService,
             issuer = IssuerAgent(
@@ -88,27 +137,66 @@ class OidvciEncryptionTest : FunSpec({
                 randomSource = RandomSource.Default
             ),
             credentialSchemes = setOf(ConstantIndex.AtomicAttribute2023),
-            requireEncryption = true, // this is important, to require encryption
-            encryptCredentialRequest = object : EncryptJweFun {
-                override suspend fun invoke(
-                    header: JweHeader,
-                    payload: String,
-                    recipientKey: JsonWebKey,
-                ): KmmResult<JweEncrypted> = KmmResult.catching {
-                    TODO("issuer fails to encrypt")
-                }
-            }
+            encryptionService = IssuerEncryptionService(
+                requireResponseEncryption = true,
+                decryptionKeyMaterial = EphemeralKeyWithoutCert(),
+                requireRequestEncryption = true, // this is important for this test
+            ),
         )
-        val requestOptions = WalletService.RequestOptions(
-            ConstantIndex.AtomicAttribute2023,
-            ConstantIndex.CredentialRepresentation.PLAIN_JWT
-        )
+
+        val requestOptions = WalletService.RequestOptions(ConstantIndex.AtomicAttribute2023, PLAIN_JWT)
         val credentialFormat = client.selectSupportedCredentialFormat(requestOptions, issuer.metadata).shouldNotBeNull()
         val scope = credentialFormat.scope.shouldNotBeNull()
         val token = getToken(scope)
-        val clientNonce = issuer.nonce().getOrThrow().clientNonce
 
-        client.createCredentialRequest(token, issuer.metadata, credentialFormat, clientNonce).getOrThrow().forEach {
+        client.createCredentialRequest(
+            tokenResponse = token,
+            metadata = issuer.metadata,
+            credentialFormat = credentialFormat,
+            clientNonce = issuer.nonce().getOrThrow().clientNonce
+        ).getOrThrow().forEach {
+            shouldThrow<OAuth2Exception.InvalidEncryptionParameters> {
+                issuer.credential(
+                    authorizationHeader = token.toHttpHeaderValue(),
+                    params = it,
+                    credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
+                ).getOrThrow()
+            }
+        }
+    }
+
+    test("issuer fails to encrypt response") {
+        issuer = CredentialIssuer(
+            authorizationService = authorizationService,
+            issuer = IssuerAgent(
+                identifier = "https://issuer.example.com".toUri(),
+                randomSource = RandomSource.Default
+            ),
+            credentialSchemes = setOf(ConstantIndex.AtomicAttribute2023),
+            encryptionService = IssuerEncryptionService(
+                requireResponseEncryption = true,
+                encryptCredentialResponse = object : EncryptJweFun {
+                    override suspend fun invoke(
+                        header: JweHeader,
+                        payload: String,
+                        recipientKey: JsonWebKey,
+                    ): KmmResult<JweEncrypted> = KmmResult.catching {
+                        TODO("issuer fails to encrypt")
+                    }
+                }
+            ),
+        )
+        val requestOptions = WalletService.RequestOptions(ConstantIndex.AtomicAttribute2023, PLAIN_JWT)
+        val credentialFormat = client.selectSupportedCredentialFormat(requestOptions, issuer.metadata).shouldNotBeNull()
+        val scope = credentialFormat.scope.shouldNotBeNull()
+        val token = getToken(scope)
+
+        client.createCredentialRequest(
+            tokenResponse = token,
+            metadata = issuer.metadata,
+            credentialFormat = credentialFormat,
+            clientNonce = issuer.nonce().getOrThrow().clientNonce
+        ).getOrThrow().forEach {
             shouldThrowAny {
                 issuer.credential(
                     authorizationHeader = token.toHttpHeaderValue(),
@@ -119,35 +207,5 @@ class OidvciEncryptionTest : FunSpec({
         }
     }
 
-    test("decrypt received credential") {
-        val requestOptions = WalletService.RequestOptions(
-            ConstantIndex.AtomicAttribute2023,
-            ConstantIndex.CredentialRepresentation.PLAIN_JWT
-        )
-        val credentialFormat =
-            client.selectSupportedCredentialFormat(requestOptions, issuer.metadata).shouldNotBeNull()
-        val scope = credentialFormat.scope.shouldNotBeNull()
-        val token = getToken(scope)
-        val clientNonce = issuer.nonce().getOrThrow().clientNonce
-
-        client.createCredentialRequest(token, issuer.metadata, credentialFormat, clientNonce).getOrThrow().forEach {
-            val credential = issuer.credential(
-                token.toHttpHeaderValue(),
-                it,
-                credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-            ).getOrThrow()
-            val serializedCredential = credential.credentials.shouldNotBeEmpty()
-                .first().credentialString.shouldNotBeNull()
-            val jwe = JweEncrypted.Companion.deserialize(serializedCredential).getOrThrow()
-            val plain = decryptJwe(jwe).getOrThrow().payload
-
-            JwsSigned.Companion.deserialize<VerifiableCredentialJws>(
-                VerifiableCredentialJws.Companion.serializer(),
-                plain,
-                vckJsonSerializer
-            ).getOrThrow()
-                .payload.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
-        }
-    }
 
 })