+++ CHANGELOG.md
@@ -56,6 +56,11 @@ Release 5.9.0 (unreleased):
    - In `TokenResponseParameters` remove `clientNonce` that has been dropped in OID4VCI draft 14
    - In `CredentialRequestParameters` deprecate `proof`, use `proofs` instead
    - Use correct error values for `unknown_credential_configuration` and `unknown_credential_identifier`
+   - In `CredentialIssuer` deprecate constructor parameters `encryptCredentialRequest`, `requireEncryption`, `supportedJweAlgorithms`, `supportedJweEncryptionAlgorithms`
+   - In `CredentialIssuer` introduce constructor parameter `encryptionService` which handles credential request decryption and credential response encryption
+   - In `WalletService` deprecate constructor parameters `requestEncryption`, `decryptionKeyMaterial`, `supportedJweAlgorithm`, `supportedJweEncryptionAlgorithm`
+   - In `WalletService` introduce constructor parameter `encryptionService` which handles credential request encryption and credential response decryption
+   - In `WalletService` add method `parseCredentialResponse` to transform the received credential response from the issuer into `StoreCredentialInput`
 
 Release 5.8.0:
  - Refactor `AuthorizationServiceStrategy`